+++ Dispatcher-DLL/src/MwDll.h
@@ -19,9 +19,9 @@ namespace desm {
 		void onStopProgramm();
 		void onStartSimulation();
 		void onStopSimulation();
-		bool setTrack(int gleisId, double von, double bis, float abstand, const std::string& name);
+		bool setTrack(int gleisId, double von, double bis, double abstand, const std::string& name);
 		bool setTrackConnection(int gleisId, int gleis1, int gleis2, double von, double bis, const std::string& name, int weiche1Id, int weiche2Id);
-		bool setSignal(int signalId, int gleisId, double position, int typ, float hoehe, float distanz, const std::string& name, int direction);
+		bool setSignal(int signalId, int gleisId, double position, int typ, double hoehe, double distanz, const std::string& name, int direction);
 		bool setBalise(int gleisId, double position, int baliseId, int direction);
 		bool setLoop(int gleisId, double positionVon, double positionBis, int baliseId);
 		bool setIsolierstoss(int gleisId, double position);
